% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-spillover.R
\name{lamp_adjacency_from_nb}
\alias{lamp_adjacency_from_nb}
\title{Build an adjacency object from a neighbour list}
\usage{
lamp_adjacency_from_nb(nb, areas, method = "supplied")
}
\arguments{
\item{nb}{An \code{spdep} neighbour list.}

\item{areas}{Area identifiers in the same order as \code{nb}.}

\item{method}{Label recorded in the object.}
}
\value{
A \code{lamp_adjacency} object.
}
\description{
Wraps an existing \code{spdep} neighbour list, for example the lattice built by
\code{\link[=lamp_simulate]{lamp_simulate()}}, in the object the spillover estimators expect.
}
\examples{
sim <- lamp_simulate(n_areas = 16, n_months = 12, design = "spillover", seed = 1)
truth <- attr(sim, "truth")
lamp_adjacency_from_nb(truth$nb, truth$lattice$area)
}
\seealso{
Other panel: 
\code{\link{lamp_adjacency}()},
\code{\link{lamp_boundaries}()},
\code{\link{lamp_changelog}()},
\code{\link{lamp_coverage}()},
\code{\link{lamp_panel}()},
\code{\link{lamp_population}()},
\code{\link{lamp_revintage}()}
}
\concept{panel}
